rename package name `org.apache.flink.table.functions.sqlFunctions` to `org.apache.flink.table.functions.sql`
